wm/sway: migrate the Sway backend to Rust

+++ src/services/window_manager_service/window_manager_service.h
@@ -38,7 +38,7 @@ window_manager_service_string_to_event(char *str) {
 typedef struct _WMWorkspace {
     gchar *name;
     gchar *output;
-    guint32 id;
+    guint64 id;
     gint32 num;
     gboolean urgent;
     gboolean focused;
